refactor: add extract-chart-series utils
